Add regression detector
